23 Add User Image Preprocessing
This commit adds functionality to preprocess user-submitted images for predictions to ensure they adhere to the input size restrictions of the current active model. Also included is a fix in the 'create_request' endpoint to convert the raw images to a format that is consistent with the rest of the application before adding it to a new Job. Co-authored by: Kaisa Arumeel <arumeel@student.chalmers.se>