#23 Refactor cache.py
This commit refactors the query_db_or_cache function inside cache.py to also handle storage and loading of the localization encoder, which is now
being returned by a function call inside this file. The encoder is also added to the return values so it can be saved with the model from the calling
function.